Fixed STA issues, CMakeLists and clang-format

+++ src/kernel/frame.cpp
@@ -7,8 +7,7 @@ namespace renderer {
 Frame::Frame(Width width, Height height)
     : width_(width),
       height_(height),
-      frame_data_(static_cast<int>(height_),
-                  std::vector<Color>(static_cast<int>(width_))) {
+      frame_data_(height_, std::vector<Color>(width_)) {
   assert(width_ > Width{0});
   assert(height_ > Height{0});
 }
@@ -23,17 +22,17 @@ Height Frame::GetHeight() const {
 
 void Frame::SetColor(Width x, Height y, Color color) {
   assert(CheckBounds(x, y));
-  frame_data_[static_cast<int>(y)][static_cast<int>(x)] = color;
+  frame_data_[y][x] = color;
 }
 
 void Frame::BlendColor(Width x, Height y, Color color) {
   assert(CheckBounds(x, y));
-  frame_data_[static_cast<int>(y)][static_cast<int>(x)].Blend(color);
+  frame_data_[y][x].Blend(color);
 }
 
 const Color& Frame::GetColor(Width x, Height y) const {
   assert(CheckBounds(x, y));
-  return frame_data_[static_cast<int>(y)][static_cast<int>(x)];
+  return frame_data_[y][x];
 }
 
 bool Frame::CheckBounds(Width x, Height y) const {
